requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 配置文件处理
PyYAML>=6.0

# 实验跟踪
wandb>=0.15.0

# 进度条
tqdm>=4.60.0

# 科学计算
scipy>=1.7.0

# 图像处理
Pillow>=8.0.0

# 日志
loguru>=0.6.0

# 工具
einops>=0.6.0